fix mingw warnings, round 3
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@7274 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/bzfs/commands.cxx
@@ -997,10 +997,10 @@ void handleShowgroupCmd(int t, const char *message)
 	line += " ";
 	itr++;
       }
-      while (line.size() > MessageLen) {
+      while (line.size() > (unsigned int)MessageLen) {
         sendMessage(ServerPlayer, t, line.substr(0, MessageLen).c_str());
         line.erase(line.begin(), line.begin() + (MessageLen - 1));
-			}
+      }
       sendMessage(ServerPlayer, t, line.c_str());
     } else {
       sendMessage(ServerPlayer, t, "There is no user by that name");
